refactor: split into clean modules, add Biome linting
