+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const NOTION_VERSION = "2026-03-11";
+export const DEFAULT_PAGE_SIZE = 100;
+export const KNOWN_AGENT_IDS = ["default", "gf_agent"] as const;
