+++ test/auth.test.ts
@@ -1,0 +1,31 @@
+import { describe, expect, it } from "vitest";
+import { getNotionApiKey } from "../src/auth.js";
+
+describe("Key isolation", () => {
+  it("default and gf_agent resolve to different API keys", () => {
+    const defaultKey = getNotionApiKey(undefined);
+    const gfKey = getNotionApiKey("gf_agent");
+    expect(defaultKey).not.toBe(gfKey);
+    expect(defaultKey.length).toBeGreaterThan(10);
+    expect(gfKey.length).toBeGreaterThan(10);
+  });
+
+  it("explicitly passing 'main' falls back to the default key", () => {
+    expect(getNotionApiKey("main")).toBe(getNotionApiKey(undefined));
+  });
+
+  it("unknown agentId falls back to default key", () => {
+    expect(getNotionApiKey("nonexistent_agent_xyz")).toBe(getNotionApiKey(undefined));
+  });
+
+  it("API keys start with the Notion secret prefix", () => {
+    const hasValidPrefix = (value: string) =>
+      value.startsWith("ntn_") || value.startsWith("secret_");
+    expect(hasValidPrefix(getNotionApiKey(undefined))).toBe(true);
+    expect(hasValidPrefix(getNotionApiKey("gf_agent"))).toBe(true);
+  });
+
+  it("repeated reads return the same default key", () => {
+    expect(getNotionApiKey(undefined)).toBe(getNotionApiKey(undefined));
+  });
+});
